refactor: streamline exports and rename page components for consistency
