+++ src/components/common/index.ts
@@ -1,0 +1,3 @@
+import ArticleItems from './ArticleItems';
+
+export { ArticleItems };
