+++ src/components/global/index.ts
@@ -1,0 +1,4 @@
+import Header from './Header';
+import Footer from './Footer';
+
+export { Header, Footer };
