+++ src/components/profile/index.ts
@@ -1,0 +1,4 @@
+import UserBanner from './UserBanner';
+import UserArticleList from './UserArticleList';
+
+export { UserBanner, UserArticleList };
